WIP: Build an abstraction for printing output, in order to allow it to be tested.
Utilizing a technique of overriding enum CodingKeys in order to Decoders a way to skip properties when initializing. In this case, we don't want a Decoder to try to initialize the printer property.

While this seems to work, it's not ideal so I'm going to continue exploring designs for the abstraction.