Completes day01, part 1
The main change is to add a protocol to support reading from the
file system in a way that allows that capability to be mocked by the
tests. The protocol has a real implementation and an in memory
implementation (just meant for tests).

This also includes the very simple algorithm to solve the exercise.